requirements.txt
fastapi
uvicorn
pydub
openai
python-dotenv
yt-dlp